+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.5.0
+
+- Support for OpenAI text to speech voices
+- Updated to latest OpenAI chat models
+
 ## 1.4.6 (2024-04-13)
 
 - Send button no longer appears in top left corner after closing a generated image for Amiga 3.X